@@ -12,9 +12,14 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
+    "babel-preset-react-native": "^1.9.0",
+    "eslint": "^2.13.0",
+    "eslint-plugin-import": "^1.9.1",
+    "eslint-plugin-jsx-a11y": "^1.5.3",
+    "eslint-plugin-react": "^5.2.2",
     "firebase": "^2.4.2",
-    "react": "^15.0.2",
-    "react-native": "^0.26.0",
+    "react": "^15.1.0",
+    "react-native": "^0.27.2",
     "react-native-vector-icons": "^2.0.3",
     "react-redux": "^4.4.5",
     "redux": "^3.5.2",
